requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
black==23.9.1
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.5
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
flake8==6.1.0
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
